convert back to int

+++ pointers_arrays_strings/100-atoi.c
@@ -40,5 +40,5 @@ int _atoi(char *s)
 	}
 
 	/* Apply negative*/
-	return (result * sign);
+	return (int)(result * sign);
 }
